requirements.txt
numpy==1.16.
matplotlib==2.1.0
protobuf==3.7.1
tensorflow-gpu==1.4.0
tqdm==4.46.0
scipy==1.2.1
scikit-learn==0.20.4
pathlib==1.0.1
llvmlite==0.30.0
numba==0.46.0
umap-learn==0.3.9